requirements.txt
# Database connectivity
psycopg2-binary==2.9.9

# HTTP requests
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Development dependencies (optional)
ipython==8.18.1
